Refactor email service to notification service, update package names, and implement notification functionality with Kafka integration
