--- src/main/java/vn/vinaacademy/email/EmailServiceApplication.java
@@ -1,23 +1,0 @@
-package vn.vinaacademy.email;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
-import org.springframework.context.annotation.ComponentScan;
-import org.springframework.context.annotation.Import;
-import vn.vinaacademy.kafka.KafkaConsumerConfig;
-
-@SpringBootApplication
-@EnableDiscoveryClient
-@ComponentScan(basePackages = {
-        "vn.vinaacademy.email",
-        "vn.vinaacademy.common"
-})
-@Import(KafkaConsumerConfig.class)
-public class EmailServiceApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(EmailServiceApplication.class, args);
-    }
-
-}
