--- src/main/java/vn/vinaacademy/email/config/AppConfig.java
@@ -1,9 +1,0 @@
-package vn.vinaacademy.email.config;
-
-import org.springframework.boot.context.properties.EnableConfigurationProperties;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-@EnableConfigurationProperties(MailProperties.class)
-public class AppConfig {
-}
