--- src/main/java/vn/vinaacademy/email/config/KafkaTopicConfig.java
@@ -1,19 +1,0 @@
-package vn.vinaacademy.email.config;
-
-import org.apache.kafka.clients.admin.NewTopic;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.kafka.config.TopicBuilder;
-
-@Configuration
-public class KafkaTopicConfig {
-    @Value("${spring.kafka.consumer.topic:email-topic}")
-    private String topicName;
-
-    @Bean
-    public NewTopic createTopic() {
-        return TopicBuilder.name(topicName)
-                .build();
-    }
-}
