--- src/main/java/vn/vinaacademy/email/config/RedisConfig.java
@@ -1,58 +1,0 @@
-package vn.vinaacademy.email.config;
-
-import lombok.RequiredArgsConstructor;
-import lombok.extern.log4j.Log4j2;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.data.redis.cache.RedisCacheConfiguration;
-import org.springframework.data.redis.cache.RedisCacheManager;
-import org.springframework.data.redis.connection.RedisConnectionFactory;
-import org.springframework.data.redis.core.RedisTemplate;
-import org.springframework.data.redis.listener.ChannelTopic;
-import org.springframework.data.redis.listener.RedisMessageListenerContainer;
-import org.springframework.data.redis.serializer.GenericJackson2JsonRedisSerializer;
-import org.springframework.data.redis.serializer.RedisSerializationContext;
-import org.springframework.data.redis.serializer.StringRedisSerializer;
-import vn.vinaacademy.email.mq.redis.EmailSubscriber;
-
-import static vn.vinaacademy.email.mq.redis.EmailQueueConstant.EMAIL_CHANNEL;
-
-@ConditionalOnProperty(name = "spring.data.redis.enabled", havingValue = "true")
-@Configuration
-@Log4j2
-@RequiredArgsConstructor
-public class RedisConfig {
-
-
-    @Bean
-    public RedisTemplate<String, Object> redisTemplate(RedisConnectionFactory redisConnectionFactory) {
-        RedisTemplate<String, Object> template = new RedisTemplate<>();
-        template.setConnectionFactory(redisConnectionFactory);
-
-        // key serializer
-        template.setKeySerializer(new StringRedisSerializer());
-        template.setHashKeySerializer(new StringRedisSerializer());
-
-        // value serializer
-        GenericJackson2JsonRedisSerializer jackson2JsonRedisSerializer = new GenericJackson2JsonRedisSerializer();
-        template.setValueSerializer(jackson2JsonRedisSerializer);
-        template.setHashValueSerializer(jackson2JsonRedisSerializer);
-
-        return template;
-    }
-
-    @Bean
-    public RedisMessageListenerContainer emailMessageContainer(RedisConnectionFactory connectionFactory, EmailSubscriber emailSubscriber) {
-        RedisMessageListenerContainer container = new RedisMessageListenerContainer();
-        container.setConnectionFactory(connectionFactory);
-        container.addMessageListener(emailSubscriber, new ChannelTopic(EMAIL_CHANNEL));
-
-        container.setErrorHandler(e -> {
-            log.error("Error in Redis message listener container", e);
-        });
-        return container;
-    }
-
-}
