--- src/main/java/vn/vinaacademy/email/dto/EmailMessage.java
@@ -1,19 +1,0 @@
-package vn.vinaacademy.email.dto;
-
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-@JsonIgnoreProperties(ignoreUnknown = true)
-public class EmailMessage {
-    private String to;
-    private String subject;
-    private String body;
-    private boolean enableHtml = false;
-}
