--- src/main/java/vn/vinaacademy/email/mq/redis/EmailQueueConstant.java
@@ -1,6 +1,0 @@
-package vn.vinaacademy.email.mq.redis;
-
-public class EmailQueueConstant {
-    public static final String EMAIL_QUEUE = "emailQueue";
-    public static final String EMAIL_CHANNEL = "emailChannel";
-}
