+++ src/main/java/vn/vinaacademy/notification/constants/NotificationAction.java
@@ -1,0 +1,7 @@
+package vn.vinaacademy.notification.constants;
+
+public class NotificationAction {
+    public static final String READ = "READ";
+    public static final String CREATE = "CREATE";
+    public static final String DELETE = "DELETE";
+}
