+++ src/main/java/vn/vinaacademy/notification/controller/NotificationController.java
@@ -1,0 +1,72 @@
+package vn.vinaacademy.notification.controller;
+
+import io.swagger.v3.oas.annotations.Operation;
+import io.swagger.v3.oas.annotations.security.SecurityRequirement;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.data.domain.Pageable;
+import org.springframework.data.domain.Sort;
+import org.springframework.web.bind.annotation.*;
+import vn.vinaacademy.common.constant.AuthConstants;
+import vn.vinaacademy.common.response.ApiResponse;
+import vn.vinaacademy.common.security.annotation.HasAnyRole;
+import vn.vinaacademy.notification.dto.NotificationDTO;
+import vn.vinaacademy.notification.entity.Notification;
+import vn.vinaacademy.notification.service.NotificationService;
+
+import java.util.UUID;
+
+@RestController
+@RequestMapping("/api/v1/notifications")
+@Slf4j
+@RequiredArgsConstructor
+@SecurityRequirement(name = "bearerAuth")
+public class NotificationController {
+
+    private final NotificationService notificationService;
+    
+    @HasAnyRole({AuthConstants.STUDENT_ROLE, AuthConstants.INSTRUCTOR_ROLE, AuthConstants.ADMIN_ROLE, AuthConstants.STAFF_ROLE})
+    @Operation(summary = "Lấy thông báo", description = "Danh sách thông báo của student")
+    @GetMapping("/paginated")
+    public ApiResponse<Page<NotificationDTO>> getUserNotificationsPaginated(
+            @RequestParam(required = false) Notification.NotificationType type,
+            @RequestParam(required = false) Boolean isRead,
+            @RequestParam(defaultValue = "0") int page,
+            @RequestParam(defaultValue = "8") int size,
+            @RequestParam(defaultValue = "createdAt") String sortBy,
+            @RequestParam(defaultValue = "desc") String direction
+    ) {
+    	
+        Sort sort = direction.equalsIgnoreCase("asc") ? Sort.by(sortBy).ascending() : Sort.by(sortBy).descending();
+        Pageable pageable = PageRequest.of(page, size, sort);
+        log.debug("get pagination page: {}, size: {}, sortBy: {}, direction: {}, type: {}, isRead: {}", page, size, sortBy, direction, type, isRead);
+        return ApiResponse.success(notificationService.getUserNotificationsPaginated(isRead, type, pageable));
+        
+    }
+    
+    @HasAnyRole({AuthConstants.STUDENT_ROLE, AuthConstants.INSTRUCTOR_ROLE, AuthConstants.ADMIN_ROLE, AuthConstants.STAFF_ROLE})
+    @Operation(summary = "Đọc thông báo", description = "Đánh dấu thông báo này đã đọc")
+    @PutMapping("/{notificationId}/read")
+    public void markAsRead(@PathVariable UUID notificationId) {
+        notificationService.markAsRead(notificationId);
+        log.debug("Mark read notification ", notificationId);
+    }
+    
+    @HasAnyRole({AuthConstants.STUDENT_ROLE, AuthConstants.INSTRUCTOR_ROLE, AuthConstants.ADMIN_ROLE, AuthConstants.STAFF_ROLE})
+    @Operation(summary = "Đọc tất cả thông báo", description = "Đánh dấu tất cả thông báo này đã đọc")
+    @PostMapping("/readall")
+    public void markAllAsRead() {
+        notificationService.markReadAll();
+        log.debug("Marked read all notifications");
+    }
+    
+    @HasAnyRole({AuthConstants.STUDENT_ROLE, AuthConstants.INSTRUCTOR_ROLE, AuthConstants.ADMIN_ROLE, AuthConstants.STAFF_ROLE})
+    @Operation(summary = "Xóa thông báo", description = "Đánh dấu thông báo này bị xóa")
+    @DeleteMapping("/{notificationId}")
+    public void deleteNotification(@PathVariable UUID notificationId) {
+        notificationService.deleteNotification(notificationId);
+        log.debug("Mark delete notification {}", notificationId);
+    }
+}
