+++ src/main/java/vn/vinaacademy/notification/dto/EmailMessage.java
@@ -1,0 +1,19 @@
+package vn.vinaacademy.notification.dto;
+
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+@Builder
+@JsonIgnoreProperties(ignoreUnknown = true)
+public class EmailMessage {
+    private String to;
+    private String subject;
+    private String body;
+    private boolean enableHtml = false;
+}
