+++ src/main/java/vn/vinaacademy/notification/dto/NotificationDTO.java
@@ -1,0 +1,26 @@
+package vn.vinaacademy.notification.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+import vn.vinaacademy.notification.entity.Notification;
+
+import java.time.LocalDateTime;
+import java.util.UUID;
+
+@Data
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+public class NotificationDTO {
+    private UUID id;
+    private String title;
+    private String content;
+    private Boolean isRead;
+    private LocalDateTime createdAt;
+    private LocalDateTime readAt;
+    private String targetUrl;
+    private Notification.NotificationType type;
+    private String email;
+}
