+++ src/main/java/vn/vinaacademy/notification/external/client/dto/Role.java
@@ -1,0 +1,16 @@
+package vn.vinaacademy.notification.external.client.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+import java.util.UUID;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+public class Role {
+    private Long id;
+    private String name;
+    private String code;
+}
