+++ src/main/java/vn/vinaacademy/notification/mq/redis/EmailQueueConstant.java
@@ -1,0 +1,6 @@
+package vn.vinaacademy.notification.mq.redis;
+
+public class EmailQueueConstant {
+    public static final String EMAIL_QUEUE = "emailQueue";
+    public static final String EMAIL_CHANNEL = "emailChannel";
+}
