--- src/test/java/vn/vinaacademy/email/EmailServiceApplicationTests.java
@@ -1,13 +1,0 @@
-package vn.vinaacademy.email;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class EmailServiceApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
