+++ src/test/java/vn/vinaacademy/notification/NotificationApplicationTests.java
@@ -1,0 +1,13 @@
+package vn.vinaacademy.notification;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class NotificationApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
